Convert buffer_rgb to integers before writing. Some transitions and presets write floats which causes exceptions when rendering the data to the network stream.
(cherry picked from commit 1f7e495c9b3c3e9f4eaa45a1f28464e2088616ed)

+++ core/networking.py
@@ -31,7 +31,8 @@ class Networking:
         """
         strand_settings = self._app.scene.get_strand_settings()
 
-        buffer_rgb = hls_to_rgb(buffer) * 255
+        # Protect against presets or transitions that write float data.
+        buffer_rgb = np.int_(hls_to_rgb(buffer) * 255)
 
         def fill_packet(intbuffer, start, end, offset, packet, swap_order=False):
             for pixel_index, pixel in enumerate(intbuffer[start:end]):
